+++ ProyectoAnime/lib/main.dart
@@ -1,5 +1,5 @@
+import 'package:ProyectoAnime/Screens/MyHomePage.dart';
 import 'package:flutter/material.dart';
-import 'package:ProyectoAnime/functions/functions.dart';
 
 void main() {
   runApp(MyApp());
@@ -14,101 +14,7 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: MyHomePage(),
+      routes: {'/one': (context) => MyHomePage()},
     );
   }
 }
-
-class MyHomePage extends StatefulWidget {
-  @override
-  MyHomePageStatus createState() => MyHomePageStatus();
-}
-
-class MyHomePageStatus extends State<MyHomePage> {
-  var busqueda;
-  var nombre;
-  int result;
-  var _lista = ["Anime", "Manga", "Persona", "Personaje"];
-  String _vista = "Seleccione una opción";
-  var _tipo;
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.black,
-        title: Image.asset(
-          "assets/iconologo.png",
-          height: 50,
-        ),
-      ),
-      body: Column(
-        children: <Widget>[
-          DropdownButton(
-            items: _lista
-                .map((String a) => DropdownMenuItem(value: a, child: Text(a)))
-                .toList(),
-            onChanged: (_value) {
-              setState(() {
-                _vista = _value;
-                _tipo = tipo(_value);
-              });
-            },
-            hint: Text(_vista),
-          ),
-          TextField(
-            decoration: InputDecoration(
-                hintText: "Busca un anime", icon: Icon(Icons.search)),
-            onSubmitted: (String str) async {
-              setState(() {
-                id(str, _tipo).then((value) {
-                  result = value;
-                });
-              });
-            },
-          ),
-          Padding(
-            padding: const EdgeInsets.all(12.0),
-            child: Text(
-              "Top 10 Mejores",
-              style: TextStyle(
-                color: Colors.red,
-                fontSize: 20.0,
-              ),
-            ),
-          ),
-          DefaultTabController(
-              length: 2,
-              child: TabBar(
-                tabs: [
-                  Tab(
-                    child: Text(
-                      "Anime",
-                      style: TextStyle(
-                        color: Colors.black,
-                        fontSize: 20.0,
-                      ),
-                    ),
-                  ),
-                  Tab(
-                    child: Text(
-                      "Manga",
-                      style: TextStyle(
-                        color: Colors.black,
-                        fontSize: 20.0,
-                      ),
-                    ),
-                  ),
-                ],
-              )),
-        ],
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () async {
-          print(result);
-        },
-      ),
-    );
-  }
-}
-
-// comentario
